refactor: handle external login via ExternalAuthController
