+++ api/Avancira.API.Tests/ExternalAuthControllerTests.cs
@@ -1,4 +1,6 @@
+using System;
 using Avancira.API.Controllers;
+using Avancira.Infrastructure.Auth;
 using FluentAssertions;
 using Microsoft.AspNetCore.Mvc;
 using Xunit;
@@ -8,14 +10,15 @@ public class ExternalAuthControllerTests
     [Theory]
     [InlineData("google")]
     [InlineData("facebook")]
-    public void ExternalLogin_WithValidProvider_Redirects(string provider)
+    public void ExternalLogin_WithValidProvider_RedirectsWithCallback(string provider)
     {
         var controller = new ExternalAuthController();
 
         var result = controller.ExternalLogin(provider);
 
+        var encodedCallback = Uri.EscapeDataString("/api/auth/external/callback");
         result.Should().BeOfType<RedirectResult>()
-            .Which.Url.Should().Be($"/api/auth/external/{provider}");
+            .Which.Url.Should().Be($"{AuthConstants.Endpoints.Authorize}?{AuthConstants.Parameters.Provider}={provider}&{AuthConstants.Parameters.RedirectUri}={encodedCallback}");
     }
 
     [Theory]
